Split client registration and track invisibles
Previously, the registration function was holding the client's
write lock for much longer than necessary, including all the welcome
messages sent after registration.
It's dangerous to hold client locks too long, because iterating
on over a channel or the server's user list in the same function
would deadlock.
In this case the welcome messages want to count invisible users,
so we really needed to release the user write lock earlier.

As a result, on_client_registered also takes a non-mut &Client now.